refactor: CardGame에서 결과 시 Map만을 반환하여, 이후 처리를 Controller에서 담당하도록 수정
