+++ src/main/java/controller/GameStatisticResponse.java
@@ -1,4 +1,4 @@
-package domain.game;
+package controller;
 
 import domain.player.Dealer;
 import domain.player.DealerCompeteResult;
@@ -8,13 +8,13 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
 
-public class GameStatistic {
+public class GameStatisticResponse {
 
     private final Dealer dealer;
     private final List<Player> players;
     private final Map<Player, DealerCompeteResult> dealerResultPerPlayer;
 
-    public GameStatistic(final Dealer dealer, final List<Player> players, final Map<Player, DealerCompeteResult> dealerResultPerPlayer) {
+    public GameStatisticResponse(final Dealer dealer, final List<Player> players, final Map<Player, DealerCompeteResult> dealerResultPerPlayer) {
         this.dealer = dealer;
         this.players = new ArrayList<>(players);
         this.dealerResultPerPlayer = dealerResultPerPlayer;
